Code refracting to flag Service
